[chore] refactoring functions into their correct places
